requirements.txt
cvxpy == 1.4.2
datetime==5.5.0
fastapi == 0.110.2
geopy == 2.4.1
h5py == 3.10.0
matplotlib == 3.8.3
motor == 3.4.0
numpy == 1.26.4
pandas == 2.2.1
pre-commit == 3.7.0
psycopg2 == 2.9.9
python-dotenv==1.0.1
requests == 2.31.0
sqlalchemy == 2.0.29
uvicorn == 0.29.0
